Some UI simplification and improvement.
Tracking is now always on, and the option has been removed.  The slider
value is only output when it changes.  Enter and Escape (and friends)
now exit with status 0 and 1, respectively.  Scripts could interpret
this as accept or abort, for example.  Space is repurposed as the
special key and tab is ignored.  Options have been reordered a bit.
Resolves #1.